[test] Extract for injection
